+++ src/test/K6/src/api/platform/events/events.js
@@ -1,7 +1,7 @@
 import http from "k6/http";
-import * as config from "../../config.js";
-import * as header from "../../buildrequestheaders.js"
-import * as support from "../../support.js";
+import * as config from "../../../config.js";
+import * as header from "../../../buildrequestheaders.js"
+import * as support from "../../../support.js";
 
 /**
  * Api call to platform events to add a cloud event
